imx8mp: ddr: Pass the dram info from SPL to U-Boot
Some DRAM vendors report the same dram information in mr[5-9] register
for the different dram sizes. As a result the U-Boot does not have an
option to get the dram size from the mr[5-8] registers. In order to
solve this problem a SHARED_DDR_INFO approach was implemented

SHARED_DDR_INFO is the area that stores the dram information detected by
the SPL code. The same address is accessible from the U-Boot code. As of now
SHARED_DDR_INFO address is used for handing over the memory info to the
U-Boot.

SHARED_DDR_INFO address is equal to the SPL_STACK address.
NOTE: the stack grows down and it never goes back to is bottom at SPL.

TBD:
This approach will be removed as soon as the handoff code gets provided
for the platform.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>

%% original patch: 0063-imx8mp-ddr-Pass-the-dram-info-from-SPL-to-U-Boot.patch